#71 make this easier for new pkg ext teams to directly use
i feel like we forgot the purpose of admiraltemplate here - people dont need to track the changes of this template as they always only care about the latest version. we should set it up so that each pkg ext team has the minimum amount of updates and clarity on how they should update.